+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add suffix prop to menuItem and submenu
+- Add prefix prop to menuItem and submenu
 
 ## [0.2.0] - 2020-05-08